Fazendo a programacao orientada e botando as funcoes nos devidos lugares
