+++ src/model/entities/contas/Cliente.java
@@ -27,9 +27,5 @@ public class Cliente extends Usuario{
         this.telemovel = telemovel;
     }
 
-    /* Metodo para verificar se o usuário e a senha estão corretos */
-    @Override
-    public boolean autenticar() {
-        return false;
-    }
+
 }
